Refactor naming db config module
