+++ bin/www
@@ -10,4 +10,6 @@ app.set("port", port);
 
 // Create HTTP server.
 const server = http.createServer(app);
+
+// Listen on provided port, on all network interfaces.
 server.listen(port);
